Add functions to use the compass: There is a thread that get the angle of the robot (angle from the North 0-360) every 500ms, we can get this angle with the function get_robot_angle and print it with print_robot_angle. This thread is define in position.c
We can also test if the compass is working with the main compassTest.c

Update of init.c and mainTest.c according to the use of the compass

Add to turn.c, the function turn_to_angle which let the robot turn to a
specific absolute angle.
TODO: write a relative one which use the compass.
This function is tested in the turnTest.c file.